requirements.txt
aiogram==3.10.0
Pillow==10.3.0
cryptography==42.0.5
google-generativeai==0.5.2
stepic==0.5.0
fastapi==0.111.0
uvicorn==0.30.1